# perf: improve config generation/loading performance