--- a/cli/commands/node.go
+++ b/cli/commands/node.go
@@ -2,7 +2,6 @@ package commands
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
@@ -11,6 +10,7 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/golang/protobuf/proto"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
@@ -135,8 +135,7 @@ func StartWunderGraphNode(n *node.Node, opts ...Option) error {
 	}
 
 	var graphConfig wgpb.WunderGraphConfiguration
-	err = json.Unmarshal(data, &graphConfig)
-	if err != nil {
+	if err := proto.Unmarshal(data, &graphConfig); err != nil {
 		log.Error("Failed to unmarshal", zap.String("filePath", configFile), zap.Error(err))
 		return errors.New("failed to unmarshal config file")
 	}