--- a/packages/sdk/src/utils/helper.ts
+++ b/packages/sdk/src/utils/helper.ts
@@ -8,3 +8,15 @@ export const isFunction = <T extends (...args: any[]) => any = (...args: any[]) 
 export const isUndefined = (v: any): v is undefined => v === UNDEFINED;
 
 export const OBJECT = Object;
+
+/**
+ * Produces a deep copy of the given argument, used structuredClone() if available
+ * @param v Value to copy
+ * @returns Deep copy of v
+ */
+export const deepClone = <T>(v: T): T => {
+	if (typeof structuredClone === 'function') {
+		return structuredClone(v);
+	}
+	return JSON.parse(JSON.stringify(v));
+};